Refactor EVM parser Async call when getting history Filter duplciate transactions on history
