+++ src/History/index.ts
@@ -1,5 +1,6 @@
 export * from './api_helpers';
 export * from './history_helpers';
 export * from './parsers';
+export * from './evmParser';
 export * from './raw_types';
 export * from './parsed_types';
